Fixed CR issues
