+++ app/i18n/ar.json
@@ -18,6 +18,7 @@
     "arriveAt": "قت الوصول",
     "now": "الان",
     "find": "حساب المسار",
+    "resetTime": "إعادة تعيين الوقت",
     "switchStations": "تبديل المحطات",
     "switchStationsHint": "اختصار لتبديل محطة البداية بمحطة الوصول"
   },
