+++ app/i18n/ru.json
@@ -21,7 +21,10 @@
     "leaveAt": "Отправление",
     "arriveAt": "Прибытие",
     "now": "Сейчас",
-    "find": "Найти маршруты"
+    "find": "Найти маршруты",
+    "resetTime": "сбросить время",
+    "switchStations": "Перевернуть станции",
+    "switchStationsHint": "Ярлык для перестановки исходной станции со станцией назначения"
   },
   "selectStation": {
     "placeholder": "Поиск остановки",
