+++ app/screens/planner/planner-screen.tsx
@@ -28,11 +28,7 @@ import { save, load } from "../../utils/storage"
 import { donateRouteIntent } from "../../utils/ios-helpers"
 import analytics from "@react-native-firebase/analytics"
 import { useFocusEffect } from "@react-navigation/native"
-
-const ResetTimeIcon = Platform.select({
-  ios: require("../../../assets/reset-time.ios.png"),
-  android: require("../../../assets/reset-time.android.png"),
-})
+import { ResetTimeButton } from "../../components/reset-time-button/reset-time-button"
 
 const { height: deviceHeight } = Dimensions.get("screen")
 
@@ -95,13 +91,6 @@ const CHANGE_DIRECTION_WRAPPER: ViewStyle = {
   zIndex: 10,
 }
 
-const RESET_TIME_ICON_STYLE: ImageStyle = {
-  height: 24,
-  width: 24,
-  resizeMode: "contain",
-  tintColor: "black",
-}
-
 // #endregion
 
 export const PlannerScreen = observer(function PlannerScreen({ navigation }: PlannerScreenProps) {
@@ -128,6 +117,11 @@ export const PlannerScreen = observer(function PlannerScreen({ navigation }: Pla
     // https://github.com/react-native-datetimepicker/datetimepicker/issues/54#issuecomment-552951685
     setDatePickerVisibility(false)
     onDateChange(date)
+  }
+
+  const onDateReset = () => {
+    HapticFeedback.trigger("impactMedium")
+    onDateChange(new Date())
   }
 
   const originData = React.useMemo(() => {
@@ -303,13 +297,7 @@ export const PlannerScreen = observer(function PlannerScreen({ navigation }: Pla
           value={formattedDate}
           style={{ marginBottom: spacing[5] }}
           onPress={() => setDatePickerVisibility(true)}
-          sideSection={
-            formattedDate !== translate("plan.now") && (
-              <TouchableOpacity hitSlop={spacing[4]} onPress={() => onDateChange(new Date())}>
-                <Image source={ResetTimeIcon} style={[RESET_TIME_ICON_STYLE]} />
-              </TouchableOpacity>
-            )
-          }
+          endSection={formattedDate !== translate("plan.now") && <ResetTimeButton onPress={onDateReset} />}
         />
 
         <DatePickerModal
